+++ task/src/main/java/com/gplucky/task/service/TaskHistoryService.java
@@ -4,7 +4,7 @@ package com.gplucky.task.service;
  * Created by ehsy_it on 2017/1/25.
  */
 public interface TaskHistoryService {
-    int insertStartTask(String type);
+    Long insertStartTask(String type);
 
     void updateFinishedTask(Long taskId);
 }
